Create SVphv_HASAUX so HVs can stop abusing SVf_OOK
Until now, the SVf_OOK flag has been set on HVs to mean that the HV
structure has an HvAUX(). This commit gives the flag its own special
name and test/set macros.

At this point, the only remaining uses of SvOOK() refer to the actual
use of the "offset" structure used on regular scalar strings.