Make OP_COREARGS an UNOP_AUX rather than SVOP
Previously it was an SVOP but one that always stored the SV internally,
not relocating it into the pad as other SVOPs usually do.

After this change, every SVOP behaves the same way; relocating its SV
onto the pad when USE_ITHREADS is in effect. That will facilitate
another future change.