Reshape struct am_table to have U8 api_ver + U16 flags
This arranges for space to put the overload API version into the
am_table directly, conveniently at the first position. We now have space
for 16 flag bits, where previously there was space for only 8.

While we're initialising the table, set the API version to 1, in
preparation for when we might have a version 2 later on.

Nothing was previously using either the flags field or the spare field,
so we can easily rename them. No flags were previously defined.