Remove code that bothers to set up or save PL_Xpv
This variable is no longer used in any perl code. It used to be used as
a temporary during some expression macros, similar to the way PL_Sv is
used. The last reference I can find to it in real code was removed in
2012, by the commit
  6702ba93464c105fae0da7b54ac21a0485c1f0b5

(It's a little late in the 5.39.x development cycle to consider removing
the line from intrpvar.h itself; but we could do that at the start of
5.41.x).