Define a new interpreter var for viral hook annotations
Currently unused by the rest of the code. Also currently called
"viralmagic" on the expectation that Hooks will eventually simply be
called Magic v2.

This commit is extracted separately simply so it can be placed at the
very start of the branch. A full rebuild is needed either side of this
commit, so by being at the beginning this makes rebase, bisect, and
other operations easier as now you don't have to `make clean`.